feat: add custom like repository

+++ src/main/java/ru/mssecondteam/reviewservice/LikeServiceImpl.java
@@ -3,6 +3,9 @@ package ru.mssecondteam.reviewservice;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import ru.mssecondteam.reviewservice.model.Review;
+
+import java.util.List;
 
 @Service
 @RequiredArgsConstructor
@@ -11,75 +14,55 @@ public class LikeServiceImpl implements LikeService {
     private final LikeRepository repository;
 
     @Override
-    public void addLike(Long reviewId, Long userId) {
-        Like like = getLikeByUserIdAndReviewId(userId, reviewId);
-        //Like dislike = getDislikeByUserIdAndReviewId(userId, reviewId);
+    public void addLike(Review review, Long userId, Boolean isPositive) {
+        Like like = getLikeByUserIdAndReviewId(userId, review.getId());
+        String name = checkLikeOrDislike(isPositive);
 
         if (like == null) {
-            Like newLike = new Like(null, userId, reviewId, true);
+            Like newLike = new Like(null, userId, review, isPositive);
             repository.save(newLike);
-            log.info("User with id '%s' add like to review with id '%s'", userId, reviewId);
-        } else if (like.getIsPositive() == false) {
+            log.info("User with id '%s' add '%s' to review with id '%s'", userId, name, review.getId());
+            System.out.println("rrr");
+        } else if (like.getIsPositive() != isPositive) {
             repository.deleteById(like.getId());
-            log.info("User with id '%s' delete dislike to review with id '%s'", userId, reviewId);
+            log.info("User with id '%s' delete '%s' to review with id '%s'", userId, name, review.getId());
         }
 
     }
 
     @Override
-    public void deleteLike(Long reviewId, Long userId) {
+    public void deleteLike(Long reviewId, Long userId, Boolean isPositive) {
         Like like = getLikeByUserIdAndReviewId(userId, reviewId);
+        String name = checkLikeOrDislike(isPositive);
 
-        if (like != null) {
+        if (like != null && like.getIsPositive() == isPositive) {
             repository.deleteById(like.getId());
-            log.info("User with id '%s' delete like to review with id '%s'", userId, reviewId);
+            log.info("User with id '%s' delete '%s' to review with id '%s'", userId, name, reviewId);
         }
 
-    }
-
-    @Override
-    public void addDislike(Long reviewId, Long userId) {
-        Like dislike = getLikeByUserIdAndReviewId(userId, reviewId);
-        //Like dislike = getDislikeByUserIdAndReviewId(userId, reviewId);
-
-        if (dislike == null) {
-            Like newLike = new Like(null, userId, reviewId, false);
-            repository.save(newLike);
-            log.info("User with id '%s' add dislike to review with id '%s'", userId, reviewId);
-        } else if (dislike.getIsPositive() == true) {
-            repository.deleteById(dislike.getId());
-            log.info("User with id '%s' delete like to review with id '%s'", userId, reviewId);
-        }
-    }
-
-    @Override
-    public void deleteDislike(Long reviewId, Long userId) {
-        Like dislike = getLikeByUserIdAndReviewId(userId, reviewId);
-
-        if (dislike != null) {
-            repository.deleteById(dislike.getId());
-            log.info("User with id '%s' delete dislike to review with id '%s'", userId, reviewId);
-        }
     }
 
     @Override
     public Like getLikeByUserIdAndReviewId(Long userId, Long reviewId) {
-        return repository.getLikeByUserIdAndReviewId(userId, reviewId);
-    }
-
-    //@Override
-    //public Like getDislikeByUserIdAndReviewId(Long userId, Long reviewId) {
-    //    return repository.getLikeByUserIdAndReviewId(userId, reviewId);
-    //}
-
-    @Override
-    public Integer getLikes(Long reviewId) {
-        return repository.getLikesCountByReviewId(reviewId, true);
-       // if ()
+        return repository.getLikeByUserIdAndReviewId(userId, reviewId).orElse(null);
     }
 
     @Override
-    public Integer getDislikes(Long reviewId) {
-        return repository.getLikesCountByReviewId(reviewId, false);
+    public LikeDto getLikesByReviewId(Long reviewId) {
+        return repository.getLikesAndDislikesByReviewId(reviewId);
     }
+
+    @Override
+    public List<LikeDto> getLikesByListReviewsId(List <Long> reviewsIds) {
+        return repository.getLikesAndDislikesByReviewsIds(reviewsIds);
+    }
+
+    private String checkLikeOrDislike(Boolean isPositive) {
+        if (isPositive) {
+            return "like";
+        } else {
+            return "dislike";
+        }
+    }
+
 }
